+++ Scripts/Items/Addons/ArcheryButteAddon.cs
@@ -235,7 +235,7 @@ namespace Server.Items
 		{
 			base.Serialize( writer );
 
-			writer.Write( (int) 0 );
+			writer.Write( 0 );
 
 			writer.Write( MinSkill );
 			writer.Write( MaxSkill );
@@ -288,7 +288,7 @@ namespace Server.Items
 		{
 			base.Serialize( writer );
 
-			writer.Write( (int) 0 ); // version
+			writer.Write( 0 ); // version
 		}
 
 		public override void Deserialize( GenericReader reader )
@@ -317,7 +317,7 @@ namespace Server.Items
 		{
 			base.Serialize( writer );
 
-			writer.Write( (int) 0 ); // version
+			writer.Write( 0 ); // version
 		}
 
 		public override void Deserialize( GenericReader reader )
