+++ Scripts/Items/Misc/Blocker.cs
@@ -51,7 +51,7 @@ namespace Server.Items
         // +2 - Hue
         // +1 - Flags
 
-        uint serial = (uint)item.Serial.Value;
+        uint serial = item.Serial.Value;
         int itemID = 0x1183;
         int amount = item.Amount;
         Point3D loc = item.Location;
