+++ Scripts/Items/Misc/LOSBlocker.cs
@@ -59,7 +59,7 @@ namespace Server.Items
         // +2 - Hue
         // +1 - Flags
 
-        uint serial = (uint)item.Serial.Value;
+        uint serial = item.Serial.Value;
         int itemID = 0x36FF;
         int amount = item.Amount;
         Point3D loc = item.Location;
