+++ Scripts/Misc/ResourceInfo.cs
@@ -563,7 +563,7 @@ namespace Server.Items
 			if ( start == CraftResource.None )
 				return 0;
 
-			return (int)(resource - start);
+			return resource - start;
 		}
 
 		/// <summary>
